refactor: move journey context from admin to ui
