--- apps/journeys-admin/src/libs/context/index.ts
@@ -1,1 +1,0 @@
-export { useJourney, JourneyProvider } from './JourneyContext'
