+++ apps/journeys-admin/tsconfig.json
@@ -13,6 +13,13 @@
     "resolveJsonModule": true,
     "isolatedModules": true
   },
-  "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx", "next-env.d.ts"],
+  "include": [
+    "**/*.ts",
+    "**/*.tsx",
+    "**/*.js",
+    "**/*.jsx",
+    "next-env.d.ts",
+    "../../libs/journeys/ui/src/libs/context/JourneyContext.spec.tsx"
+  ],
   "exclude": ["node_modules"]
 }
