+++ libs/journeys/ui/src/libs/index.ts
@@ -29,7 +29,14 @@ export {
   previousBlocksVar,
   treeBlocksVar
 } from './useBlocks/blocks'
-export { useEditor, EditorProvider, ActiveTab, ActiveFab } from './context'
+export {
+  useEditor,
+  EditorProvider,
+  ActiveTab,
+  ActiveFab,
+  useJourney,
+  JourneyProvider
+} from './context'
 export type {
   SetSelectedStepAction,
   SetSelectedBlockByIdAction
